Push and Pull
Player now enhanced with the ability to swap between a normal sprite and one with their arms up.
Variable Weight obstacles are easily doable, and demoed in the scene.
Still playing around with how to do more complex pulling, current implementation has significant issues - trying to do it via parenting, but seem unable to get the player to drop it when they put their arms down