Multiplayer
Multiplayer is, in theory, implemented. The gamemanager now detects if there's one or two controllers detected. If it's 2, then it activates player 2. The playercontroller script now includes a toggle between player 1 & 2; this is necessary as they use different joysticks for input.
ALSO: Player 2 can only be controlled from a controller's dpad, haven't set up the inputs for kb or joystick.

NOTE!: I'M UNABLE TO TEST THIS DUE TO ONLY HAVING ONE CONTROLLER